create cli module, move util.run -> cli.run
