+++ examples/burgers.py
@@ -1,6 +1,6 @@
 #! /usr/bin/python3
 
-from nutils import mesh, util, log, function, plot, debug, _
+from nutils import mesh, cli, log, function, plot, debug, _
 import numpy
 
 
@@ -86,4 +86,4 @@ def unittest():
 
 
 if __name__ == '__main__':
-  util.run( main, unittest )
+  cli.run( main, unittest )
