+++ examples/finitestrain.py
@@ -67,4 +67,4 @@ def unittest():
 
 
 if __name__ == '__main__':
-  util.run( main, unittest )
+  cli.run( main, unittest )
