+++ nutils/__init__.py
@@ -9,4 +9,4 @@ version = '3beta'
 _ = numpy.newaxis
 __all__ = [ '_', 'numpy', 'core', 'numeric', 'element', 'function',
   'mesh', 'plot', 'library', 'topology', 'util', 'matrix', 'parallel', 'log',
-  'debug', 'cache', 'transform', 'rational', 'index', 'model' ]
+  'debug', 'cache', 'transform', 'rational', 'index', 'model', 'cli' ]
